hospital-root/master_data_service/requirements.txt
grpcio
grpcio-tools
sqlalchemy[asyncio]
asyncpg
pydantic-settings
pyjwt
confluent-kafka
alembic
